+++ src/Config_test.hs
@@ -69,7 +69,7 @@ test_pickModule = do
         Just (Just "p2", "Z.M")
 
 test_formatGroups = do
-    let f config imports = lines $ Config.formatGroups
+    let f config imports = lines $ Config.formatGroups Config.Standard
             (Config._order (parseConfig config))
             (Testing.expectRight (parse (unlines imports)))
     equal (f [] []) []
@@ -115,6 +115,14 @@ test_formatGroups = do
         , "import qualified A"
         ]
 
+test_showImportLine = do
+    let f = fmap (Config.showImportDecl style . Types.importDecl . head)
+            . parse
+        style = Config.Custom $ Config.PPConfig
+            { _leaveSpaceForQualified = True
+            }
+    -- pprint $ fmap (Types.importDecl . head) . parse $ "import A.B as C (x)"
+    equal (f "import A.B as C (x)") $ Right "import           A.B as C (x)"
 
 -- * util